backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
pydantic-settings==2.7.0
pydantic[email]==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
jinja2==3.1.4
apscheduler==3.10.4